ia-agent-tickets-v1/agent/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
